requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
sqlite-utils
